feat: add Message entity and ChatProvider to manage messages
Add Message entity to represent a chat message and ChatProvider to manage messages. The ChatProvider has a list of messages and a sendMessage method to add a new message to the list.

Also, update the ChatScreen to use the ChatProvider and display the messages. The MyMessageBubble widget is updated to display the message text from the Message entity.

Finally, add the provider package as a dependency to the pubspec.yaml file.